Create new `sources` app, and move source-related views, tests, etc. (everything besides models) from `images` app to `sources`.
