+++ project/images/utils.py
@@ -1,6 +1,4 @@
-from collections import OrderedDict
 import datetime
-import math
 import random
 
 from django.conf import settings
@@ -9,7 +7,7 @@ from django.db.models import Q
 from accounts.utils import get_alleviate_user
 from annotations.model_utils import AnnotationAreaUtils
 from .model_utils import PointGen
-from .models import Source, Point, Image, Metadata
+from .models import Image, Metadata, Point
 
 
 def _get_next_images_queryset(current_image, image_queryset):
@@ -126,46 +124,11 @@ def get_image_order_placement(current_image, image_queryset):
     return prev_images.count() + 1
 
 
-def metadata_field_names_to_labels(source):
-    """
-    Get an OrderedDict of Metadata field names to field labels.
-    e.g. 'photo_date': "Date", 'aux1': "Site", 'camera': "Camera", ...
-    """
-    d = OrderedDict(
-        (field_name, Metadata._meta.get_field(field_name).verbose_name)
-        for field_name
-        in Metadata.EDIT_FORM_FIELDS
-    )
-    # Instead of "Aux1" for field aux1, use the source specified label,
-    # e.g. "Site"
-    for num, aux_field_name in enumerate(get_aux_field_names(), 1):
-        d[aux_field_name] = get_aux_label(source, num)
-    return d
-
-
 def metadata_obj_to_dict(metadata):
     """
     Go from Metadata DB object to metadata dict.
     """
     return dict((k, getattr(metadata, k)) for k in Metadata.EDIT_FORM_FIELDS)
-
-
-def aux_label_name_collisions(source):
-    """
-    See if the given source's auxiliary metadata field labels have
-    name collisions with (1) built-in metadata fields, or (2) each other.
-    Return a list of the names which appear more than once.
-    Comparisons are case insensitive.
-    """
-    field_names_to_labels = metadata_field_names_to_labels(source)
-    field_labels_lower = [
-        label.lower() for label in field_names_to_labels.values()]
-
-    dupe_labels = [
-        label for label in field_labels_lower
-        if field_labels_lower.count(label) > 1
-    ]
-    return dupe_labels
 
 
 def delete_images(image_queryset):
@@ -371,12 +334,6 @@ def generate_points(img, usesourcemethod=True):
               point_number=new_point['point_number'],
               image=img,
         ).save()
-
-
-def filter_out_test_sources(source_queryset):
-    for possible_test_substring in settings.LIKELY_TEST_SOURCE_NAMES:
-        source_queryset = source_queryset.exclude(name__icontains=possible_test_substring)
-    return source_queryset
 
 
 def get_carousel_images():
