+++ project/jobs/templates/jobs/source_job_list.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 
 {% block title %}Jobs for {{ source.name }} | CoralNet{% endblock %}
