+++ project/labels/templates/labels/labelset_edit.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% load common_tags %}
 
 {% block title %}Edit labelset entries - {{ source.name }} - CoralNet{% endblock %}
