+++ project/labels/templates/labels/labelset_main.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% load guardian_tags %}
 
 {% block title %}Labelset for {{ source.name }} - CoralNet{% endblock %}
