+++ project/labels/templates/labels/labelset_required.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 
 {% block title %}No labelset | CoralNet{% endblock %}
