+++ project/sources/templates/sources/invites_manage.html
