+++ project/sources/templates/sources/source_invite.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 
 {% block title %}Administer Source | CoralNet{% endblock %}
