+++ project/sources/templates/sources/source_main.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% load guardian_tags %}
 {% load thumbnail %}
 {% load markdown_to_html %}
