+++ project/upload/templates/upload/upload_images.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% load thumbnail %}
 {% load common_tags %}
