+++ project/vision_backend/templates/vision_backend/backend_main.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 
 {% block js-includes %}
   {% include "javascript-includes.html" with jquery_ui="yes" %}
