+++ project/visualization/templates/visualization/browse_images.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% load async_media_batch_key thumbnail_async from async_media_tags %}
 {% load jsonify from common_tags %}
