+++ project/visualization/templates/visualization/browse_patches.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% load get_obj_perms from guardian_tags %}
 {% load async_media_batch_key patch_async from async_media_tags %}
