+++ project/visualization/templates/visualization/edit_metadata.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% load thumbnail %}
 {% load guardian_tags %}
 {% load common_tags %}
