+++ project/visualization/templates/visualization/statistics.html
@@ -1,4 +1,4 @@
-{% extends "images/base_source.html" %}
+{% extends "sources/base_source.html" %}
 {% block title %}{{ source.name }} - Statistics | CoralNet{% endblock %}
 
 {% block css-code %}
